COMP: fix type warning
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=10506
git-svn-id: http://svn.slicer.org/Slicer4/trunk@10506 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkEventBroker.cxx
@@ -467,7 +467,7 @@ vtkObservation *vtkEventBroker::GetNthObservation ( int n )
   ObjectToObservationVectorMap::iterator iter; 
   for(iter=this->SubjectMap.begin(); iter != this->SubjectMap.end(); iter++)  
     {
-    if ( n < count + iter->second.size())
+    if ( n < count + (int)iter->second.size())
       {
       return (iter->second[n-count]);
       }
